Split home shell route
